backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.4
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
pandas==2.2.3
openpyxl==3.1.5
